third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/inject-stylesheet-expected.txt

CONSOLE MESSAGE: line 10: iframe loaded
Tests that injected user stylesheets are reflected in the Styles pane.


Main frame style:
background-attachment: scroll;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-clip: border-box;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-color: rgb(0, 0, 255);
    blue - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-image: none;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-origin: padding-box;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-position-x: 0%;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-position-y: 0%;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-repeat-x: ;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-repeat-y: ;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
background-size: auto;
    initial - #main inject-stylesheet.html:5 -> inject-stylesheet.html:5:10
border-bottom-style: solid;
    solid - #main injected stylesheet
border-left-style: solid;
    solid - #main injected stylesheet
border-right-style: solid;
    solid - #main injected stylesheet
border-top-style: solid;
    solid - #main injected stylesheet
color: rgb(255, 0, 0);
    red - #main injected stylesheet
display: block;
    block - div user agent stylesheet
-webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==) 100% / 1 / 0px stretch;
    url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==) - #main injected stylesheet
[expanded] 
element.style { ()

[expanded] 
#main { (inject-stylesheet.html:5 -> inject-stylesheet.html:5:10)
    background: blue;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: blue;

[expanded] 
#main { (injected stylesheet)
    -webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==);
    color: red;
    border-style: solid;
        border-top-style: solid;
        border-right-style: solid;
        border-bottom-style: solid;
        border-left-style: solid;

[expanded] 
div { (user agent stylesheet)
    display: block;

iframe style:
background-attachment: scroll;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-clip: border-box;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-color: rgb(255, 0, 0);
    OVERLOADED green - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    red - #iframeBody injected stylesheet
background-image: none;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-origin: padding-box;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-position-x: 0%;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-position-y: 0%;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-repeat-x: ;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-repeat-y: ;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
background-size: auto;
    OVERLOADED initial - body inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9
    initial - #iframeBody injected stylesheet
display: block;
    block - body user agent stylesheet
margin-bottom: 8px;
    8px - body user agent stylesheet
margin-left: 8px;
    8px - body user agent stylesheet
margin-right: 8px;
    8px - body user agent stylesheet
margin-top: 8px;
    8px - body user agent stylesheet
[expanded] 
element.style { ()

[expanded] 
body { (inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:9)
/-- overloaded --/     background: green;
    /-- overloaded --/     background-image: initial;
    /-- overloaded --/     background-position-x: initial;
    /-- overloaded --/     background-position-y: initial;
    /-- overloaded --/     background-size: initial;
    /-- overloaded --/     background-repeat-x: initial;
    /-- overloaded --/     background-repeat-y: initial;
    /-- overloaded --/     background-attachment: initial;
    /-- overloaded --/     background-origin: initial;
    /-- overloaded --/     background-clip: initial;
    /-- overloaded --/     background-color: green;

[expanded] 
#iframeBody { (injected stylesheet)
    background: red !important;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: red;

[expanded] 
body { (user agent stylesheet)
    display: block;
    margin: 8px;
        margin-top: 8px;
        margin-right: 8px;
        margin-bottom: 8px;
        margin-left: 8px;
